requirements.txt
streamlit==1.3.1
numpy
pandas
opencv-python-headless
Pillow==8.3.2
temp
mediapipe==0.8.7
bokeh==2.4.1